# LinearStretching enhancements