--- a/src/algorithms/linear_stretching.jl
+++ b/src/algorithms/linear_stretching.jl
@@ -2,7 +2,7 @@
 """
 ```
     LinearStretching <: AbstractHistogramAdjustmentAlgorithm
-    LinearStretching(; minval = 0, maxval = 1)
+    LinearStretching(; minval = 0, maxval = 1, [src_minval], [src_maxval])
 
     adjust_histogram([T,] img, f::LinearStretching)
     adjust_histogram!([out,] img, f::LinearStretching)
@@ -44,10 +44,16 @@ type as the input.
 If minval and maxval are specified then intensities are mapped to the range
 [`minval`, `maxval`]. The default values are 0 and 1.
 
+## Choices for `src_minval` and `src_maxval`
+
+`src_minval` and `src_maxval` specifies the intensity range of input image. By default,
+the values are `extrema(img)` (finite). If custom values are provided, the output
+intensity value will be clamped to range `(minval, maxval)` if it exceeds that.
+
 # Example
 
 ```julia
-using ImageContrastAdjustment, ImageView, TestImages
+using ImageContrastAdjustment, TestImages
 
 img = testimage("mandril_gray")
 imgo = adjust_histogram(img, LinearStretching(minval = 0, maxval = 1))
@@ -59,24 +65,51 @@ imgo = adjust_histogram(img, LinearStretching(minval = 0, maxval = 1))
 
 """
 @with_kw struct LinearStretching{T₁ <: Union{Real,AbstractGray},
-                                    T₂ <: Union{Real,AbstractGray}} <: AbstractHistogramAdjustmentAlgorithm
-    minval::T₁ = 0.0
-    maxval::T₂ = 1.0
+                                 T₂ <: Union{Real,AbstractGray},
+                                 T₃ <: Union{Nothing, Real,AbstractGray},
+                                 T₄ <: Union{Nothing, Real,AbstractGray}} <: AbstractHistogramAdjustmentAlgorithm
+    minval::T₁     = 0.0
+    maxval::T₂     = 1.0
+    src_minval::T₃ = nothing
+    src_maxval::T₄ = nothing
 end
 
 function (f::LinearStretching)(out::GenericGrayImage, img::GenericGrayImage)
-    src_minval = minfinite(img)
-    src_maxval = maxfinite(img)
+    img_min, img_max = minfinite(img), maxfinite(img)
+    src_minval = isnothing(f.src_minval) ? img_min : f.src_minval
+    src_maxval = isnothing(f.src_maxval) ? img_max : f.src_maxval
     T = eltype(out)
-    out .= img
-    map!(out,out) do val
+
+    # r * x - o is equivalent to (x-A) * ((b-a)/(B-A)) + a
+    # precalculate these and make inner loop contains only multiplication and addition
+    # to get better performance
+    r = (f.maxval - f.minval) / (src_maxval - src_minval)
+    o = (src_minval*f.maxval - src_maxval*f.minval) / (src_maxval - src_minval)
+
+    if 1 ≈ r && 0 ≈ o
+        # when image intensity is already adjusted, there's no need to do it again
+        # it's a trivial but common case in practice
+        out .= img
+        return out
+    end
+
+    # In most cases, we don't need to clamp the output
+    # this is only used when user specifies custom `(src_minval, src_maxval)`
+    out_minval = r * img_min - o
+    out_maxval = r * img_max - o
+    do_clamp = (out_minval < f.minval) || (out_maxval > f.maxval)
+
+    @inbounds @simd for p in eachindex(img)
+        val = img[p]
         if isnan(val)
-            return val
+            out[p] = val
         else
-            newval = linear_stretch(val, src_minval, src_maxval, f.minval, f.maxval)
-            return  T <: Integer ? round(Int, newval ) : newval
+            newval = r * val - o
+            do_clamp && (newval = clamp(newval, f.minval, f.maxval))
+            out[p] = T <: Integer ? round(Int, newval) : newval
         end
     end
+    out
 end
 
 function (f::LinearStretching)(out::AbstractArray{<:Color3}, img::AbstractArray{<:Color3})