It's just a prototype for #27 , does this usage look good to you? 

```julia
img = float.(testimage("mandril_gray"))
imgo = adjust_histogram(img, LinearStretching(src_minval=0.1, src_maxval=0.9))
```

Note that this implementation doesn't always work for `N0f8` types, there're two ways to work this around: clamp01 or promote storage type. I'm not sure which is better.

cc: @zygmuntszpak @bjarthur