@bjarthur 2020-03-17
great!  in what cases doesn't it work for N0f8?