@johnnychen94 2020-03-19
> in what cases doesn't it work for N0f8?

When the real output value exceeds the range specified by `[minval, maxval]`, which by default is `[0, 1]`. For example, when `src_minval > finitemin(img)` or `src_maxval < finitemax(img)` it raises following error:

```
julia> imgo = adjust_histogram(img, LinearStretching(src_minval=0.1, src_maxval=0.9))
ERROR: ArgumentError: element type FixedPointNumbers.Normed{UInt8,8} is an 8-bit type representing 256 values from 0.0 to 1.0,
  but the values (-0.0024509803921568714,) do not lie within this range.
  See the READMEs for FixedPointNumbers and ColorTypes for more information.
```

Paremeters `(src_minval, src_maxval)` and `(minval, maxval)` are a little bit mutually exclusive for N0f8 types.