@bjarthur 2020-03-19
this issue should exist for all fixed-point types not just N0f8 IIUC.

`imadjustintensity` clamped the output to [0,1].  this precedent is/was intuitive to me.

however, i see your point.  give the user the flexibility to stretch outside [0,1] and then clamp themselves if that is what they desire.  and if they stretch into a type which can't support values outside [0,1], then either throw an error or promote.

maybe make clamping easy for them by having a `clamp=false` keyword arg?