@johnnychen94 2020-03-19
Changes:

* performance tweak
* post `clamp` wrt (minval, maxval)
* warn user about the range issue related to `FixedPoint` with suggestions. (Not a fix)

Performance tweak get ~10x speed up:

```julia
julia> img = testimage("mandril_gray");
julia> alg = LinearStretching();

# master
julia> @benchmark adjust_histogram($img, $alg)
BenchmarkTools.Trial:
  memory estimate:  256.19 KiB
  allocs estimate:  3
  --------------
  minimum time:     4.473 ms (0.00% GC)
  median time:      4.739 ms (0.00% GC)
  mean time:        4.804 ms (0.03% GC)
  maximum time:     7.016 ms (0.00% GC)
  --------------
  samples:          1041
  evals/sample:     1

# this PR
julia> @benchmark adjust_histogram($img, $alg)
BenchmarkTools.Trial:
  memory estimate:  256.14 KiB
  allocs estimate:  2
  --------------
  minimum time:     493.738 μs (0.00% GC)
  median time:      535.095 μs (0.00% GC)
  mean time:        550.216 μs (0.36% GC)
  maximum time:     2.837 ms (80.10% GC)
  --------------
  samples:          9068
  evals/sample:     1
```