@johnnychen94 2020-03-21
Hmmm, `Gray{Float32}` performance regression on `Gray{Float32}` is really unexpected 🤔