@johnnychen94 2020-03-21
Benchmark after 65c7c22:

* the N0f8 version is faster
* float version isn't changed
* performances are almost the same with/without `clamp`

```julia
julia> img = testimage("mandril_gray");

julia> alg = LinearStretching();
julia> @btime adjust_histogram($img, $alg);
  265.302 μs (6 allocations: 1.25 MiB)
julia> @btime adjust_histogram($(float.(img)), $alg);
  909.417 μs (2 allocations: 1.00 MiB)
julia> @btime adjust_histogram($(Float32.(img)), $alg);
  879.758 μs (2 allocations: 1.00 MiB)

julia> alg = LinearStretching((0.1f0, 0.9f0));
julia> @btime adjust_histogram($img, $alg);
  275.747 μs (6 allocations: 1.25 MiB)
julia> @btime adjust_histogram($(float.(img)), $alg);
  882.314 μs (2 allocations: 1.00 MiB)
julia> @btime adjust_histogram($(Float32.(img)), $alg);
  880.484 μs (2 allocations: 1.00 MiB)
```