@johnnychen94 2020-03-24
Summary of this PR:

* expose `src_minval` and `src_maxval`
* deprecate `minval`/`maxval` in favor of `dst_minval`/`dst_maxval`
* correctly handles the clamp on output intensity
* expose `no_clamp` option to disable the clamp to `[dst_minval, dst_maxval]`: `FixedPoint` inputs are still clamped wrt `[typemin(T), typemax(T)]` to avoid ArgumentError
* performance tweak: 2x-20x performance boost

I plan to merge this in two days if there are no further comments