@bjarthur 2020-03-31
indeed, great work!

sorry i didn't think to suggest these earlier, but two things:

1. it would be convenient to be able to specify a vector of length 2, in addition to a tuple, as the {src,dst}_{min,max}val .  this signature would also be consistent with how the `dims` keyword argument on many julia functions works (e.g. `mean`).

2. how about `clamp=true` instead of `no_clamp=false`?  i know it might seem spurious, but it is not only three fewer letters, but avoids a double negative.  the latter in particular makes me have to think a little harder to not get the sign flipped.

i can submit a PR if there is a consensus about either one or the other or both.