@johnnychen94 2020-03-31
Accepting vector seems like an anti-pattern since tuple is faster than vector. A permanent deprecation or a helpful error message(ref: https://github.com/JuliaLang/julia/pull/35094) might be helpful in this case.

I'm okay to rename it to `clamp`, but perhaps `do_clamp` is more intuitive.