@zygmuntszpak 2020-04-01
@bjarthur Could you please elaborate on how specifying a vector rather than a tuple is more convenient? I would prefer not to support two different ways of specifying the same thing. 

`do_clamp` sounds odd to me since I read `clamp` as an action (verb) already, and so the `do` seems unnecessary.  Hence, I would probably go with `clamp`. An alternative, which I anticipate people not to like because it is longer, is to have something like `clamp_result = true`.