@bjarthur 2020-04-01
i believe in julia it is standard to accept both 2-tuples and 2-vectors when specifying the dimensions along which to perform an algebraic operation.  so for example with `mean`:

```
julia> using Statistics

julia> A=rand(3,4,5);

julia> mean(A, dims=(1,3))
1×4×1 Array{Float64,3}:
[:, :, 1] =
 0.394565  0.477098  0.5762  0.563391

julia> mean(A, dims=[1,3])
1×4×1 Array{Float64,3}:
[:, :, 1] =
 0.394565  0.477098  0.5762  0.563391
```

similarly for `quantile` even:

```
julia> V=rand(100);

julia> quantile(V, (0.1,0.9))
(0.07442527144546038, 0.9287217120956874)

julia> quantile(V, [0.1,0.9])
2-element Array{Float64,1}:
 0.07442527144546038
 0.9287217120956874
```

with LinearStretching in some cases the min/max vals you want to use might be calculated and stored in a vector and it is a hassle for the user to `tuple` them.  particularly when supporting vectors i think is as simple as adding the following method:

```
LinearStretching(rangemap::Pair{Vector, Vector}; no_clamp=false) =
    LinearStretching(rangemap.first..., rangemap.second..., nothing, nothing, no_clamp)
```