@johnnychen94 2020-04-02
> with LinearStretching in some cases the min/max vals you want to use might be calculated and stored in a vector and it is a hassle for the user to tuple them.

That could be quite a fair use case for me. However, I'm not excited to add this to the codebase; I kind of prefer to distinguish tuple and vector and to not mixing them.