@zygmuntszpak 2020-04-02
@bjarthur Perhaps you could open an issue on the Images repo regarding your suggestion of also allowing one to pass a vector instead of a tuple. If we do this here, then I would like it to become a consensus for the Images ecosystem. I want to avoid a situation where some key functions do this kind of conversion automatically, and others not.